minimig: read the full 16-bit ext-bus status word
Both status readers fell back to a second word cast through uint8_t,
which discards bits 8 and above: the akiko request, sector-request and
rx-busy flags, the CDDA request, and the new CDTV NVRAM dirty flag. The
fallback never fired in practice because the status word carried a
constant 0xE tag in its top nibble, so the first read was always
non-zero. Removing that tag to free flag bits removes the accident that
was keeping the cast harmless, so drop the cast instead of relying on a
magic value to steer around it.

cdtv_read_status_full was a byte-for-byte duplicate of cdtv_read_status
that existed only to avoid the same truncation; with the cast gone the
two are identical, so fold it away.

+++ support/minimig/akiko_cd32.cpp
@@ -224,7 +224,7 @@ static uint16_t akiko_read_status(void)
 	uint16_t res;
 	EnableIO();
 	res = spi_w(AKIKO_STATUS_CMD);
-	if (!res) res = (uint8_t)spi_w(0);
+	if (!res) res = spi_w(0);
 	DisableIO();
 	return res;
 }
